rename IDownloadAsyncDecorator
